backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
sqlalchemy==2.0.46
aiosqlite==0.20.0
watchdog==4.0.0
python-multipart==0.0.9
python-dateutil==2.8.2
lxml==6.0.2
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-httpx==0.30.0
reportlab==4.2.0
openpyxl==3.1.2
aiofiles==23.2.1
rich==13.7.1
loguru==0.7.2
